refactor supCon.py
move train function to utils/train.py
